requirements.txt
# Core dependencies
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.8.2
email-validator==1.1.3
python-dotenv==0.19.0

# Database
sqlalchemy==1.4.23
aiosqlite==0.17.0
alembic==1.7.1

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.19.0
pytest-cov==2.12.1
pytest-mock==3.6.1

# Security
cryptography==3.4.8
fastapi-limiter==0.1.4
bcrypt==3.2.0

# Async utilities
aiofiles==0.7.0
asyncio==3.4.3
aiohttp==3.7.4

# Logging and monitoring
loguru==0.5.3
structlog==21.1.0
prometheus-client==0.11.0

# Data processing
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2

# Documentation
mkdocs==1.2.3
mkdocs-material==7.3.0

# Development tools
black==21.7b0
flake8==3.9.2
mypy==0.910
isort==5.9.3
pre-commit==2.15.0

# Optional dependencies
redis==3.5.3  # For rate limiting
requests==2.26.0  # For external API calls
beautifulsoup4==4.9.3  # For web scraping
python-dateutil==2.8.2
